docs/requirements.txt
flask<3
flask-jwt-extended<5
PyJWT[crypto]<3
Flask-SQLAlchemy<3
SQLAlchemy<2
psycopg2-binary>=2.8,<3
flask-serialize>=2.0.3,<3
pytest>=7.1.2,<8
Flask-Alembic>=2.0.1,<3
flask-rest-api>=0.17.1,<1
werkzeug==2.0.3
alembic~=1.8.1
Pillow>=9.2.0,<10
Flask-Pydantic>=0.9.0,<1
python-dotenv<1
